dtype='int64' for get_dummies.

+++ notebooks/solutions/titanic-age-dropna.py
@@ -5,7 +5,7 @@ train_df = age_non_null_train_df
 
 # Choose features and lables
 features = ["Pclass", "Sex", "SibSp", "Parch", "Age"]
-X = pd.get_dummies(train_df[features], drop_first=True)
+X = pd.get_dummies(train_df[features], dtype='int64', drop_first=True)
 y = train_df['Survived']
 
 # Split data into training and test data
